+++ packages/axe-core-cucumber/e2e/capybara/Gemfile
@@ -3,5 +3,7 @@ source "https://rubygems.org"
 gem "cucumber"
 gem "capybara"
 gem "selenium-webdriver"
+
+gem "axe-core-api", :path => "../../../axe-core-api"
 gem "axe-core-capybara", :path => "../../../axe-core-capybara"
 gem "axe-core-cucumber", :path => "../../"
