+++ packages/axe-core-cucumber/e2e/capybara/features/support/env.rb
@@ -3,7 +3,7 @@ require "capybara/cucumber"
 require "axe-capybara"
 # Requiring 'axe/cucumber/step_definitions' makes all of the axe-cucumber step
 # definitions available to be used directly in your cucumber features.
-require "axe-cucumber-step-definitions"
+require "axe-core-cucumber-step-definitions"
 # configure `AxeCapybara`
-@page = AxeCapybara.configure("firefox") do |c|
+@page = AxeCapybara.configure(:firefox) do |c|
 end
