+++ packages/axe-core-cucumber/lib/axe-core-cucumber-step-definitions.rb
@@ -1,0 +1,3 @@
+require_relative "./axe-core-cucumber"
+
+Then AxeCucumber::Step::REGEX, :assert_accessibility, on: -> { AxeCucumber::Step.create_for self }
