--- packages/axe-core-cucumber/lib/axe-cucumber-step-definitions.rb
@@ -1,7 +1,0 @@
-require_relative "./axe-cucumber"
-
-Then AxeCucumber::Step::REGEX,
-     :assert_accessibility,
-     on: -> {
-       AxeCucumber::Step.create_for self
-     }
