+++ packages/axe-core-cucumber/spec/step_regex_spec.rb
@@ -1,6 +1,5 @@
 require "spec_helper"
-
-require_relative "../lib/axe-cucumber"
+require_relative "../lib/axe-core-cucumber"
 
 module CustomMatchers
   class Match < RSpec::Matchers::BuiltIn::Match
@@ -185,64 +184,43 @@ module AxeCucumber
         # minimum
         it {
           is_expected.to matchy(
-            'the page should be accessible
-            within "foo"
-            excluding "bar"
-            according to: tag
-            checking: run
-            skipping: skip
-            with options: qux'
-          )
-            .capturing(negate: be_falsey,
-                       inclusion: "foo",
-                       exclusion: "bar",
-                       tags: "tag",
-                       run_rules: "run",
-                       skip_rules: "skip",
-                       run_only: be_falsey,
-                       options: "qux")
+            'the page should be accessible within "foo" excluding "bar" according to: tag checking: run skipping: skip with options: qux'
+          ).capturing(negate: be_falsey,
+                      inclusion: "foo",
+                      exclusion: "bar",
+                      tags: "tag",
+                      run_rules: "run",
+                      skip_rules: "skip",
+                      run_only: be_falsey,
+                      options: "qux")
         }
 
         # using semicolons
         it {
           is_expected.to matchy(
-            'the page should be accessible
-            within "foo";
-            excluding "bar";
-            according to: tag;
-            checking: run;
-            skipping: skip;
-            with options: qux'
-          )
-            .capturing(negate: be_falsey,
-                       inclusion: "foo",
-                       exclusion: "bar",
-                       tags: "tag",
-                       run_rules: "run",
-                       skip_rules: "skip",
-                       run_only: be_falsey,
-                       options: "qux")
+            'the page should be accessible within "foo"; excluding "bar"; according to: tag; checking: run; skipping: skip; with options: qux'
+          ).capturing(negate: be_falsey,
+                      inclusion: "foo",
+                      exclusion: "bar",
+                      tags: "tag",
+                      run_rules: "run",
+                      skip_rules: "skip",
+                      run_only: be_falsey,
+                      options: "qux")
         }
 
         # using conjunctions
         it {
           is_expected.to matchy(
-            'the page should be accessible
-                            within "foo" but
-                            excluding "bar"
-                            according to: tag and
-                            checking: run but
-                            skipping: skip
-                            with options: qux'
-          )
-            .capturing(negate: be_falsey,
-                       inclusion: "foo",
-                       exclusion: "bar",
-                       tags: "tag",
-                       run_rules: "run",
-                       skip_rules: "skip",
-                       run_only: be_falsey,
-                       options: "qux")
+            'the page should be accessible within "foo" but excluding "bar" according to: tag and checking: run but skipping: skip with options: qux'
+          ).capturing(negate: be_falsey,
+                      inclusion: "foo",
+                      exclusion: "bar",
+                      tags: "tag",
+                      run_rules: "run",
+                      skip_rules: "skip",
+                      run_only: be_falsey,
+                      options: "qux")
         }
       end
     end
